+++ apps/meteor/app/models/client/models/CachedChatRoom.ts
@@ -1,7 +1,7 @@
 import type { IOmnichannelRoom, IRoom, IRoomWithRetentionPolicy } from '@rocket.chat/core-typings';
 import { DEFAULT_SLA_CONFIG, LivechatPriorityWeight } from '@rocket.chat/core-typings';
 
-import { CachedCollection } from '../../../ui-cached-collection/client/models/CachedCollection';
+import { CachedCollection } from '../../../../client/lib/cachedCollections/CachedCollection';
 import { CachedChatSubscription } from './CachedChatSubscription';
 
 class CachedChatRoom extends CachedCollection<IRoom> {
