+++ apps/meteor/app/models/client/models/CachedChatSubscription.ts
@@ -2,7 +2,7 @@ import type { IOmnichannelRoom, IRoomWithRetentionPolicy, ISubscription } from '
 import { DEFAULT_SLA_CONFIG, LivechatPriorityWeight } from '@rocket.chat/core-typings';
 import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 
-import { CachedCollection } from '../../../ui-cached-collection/client/models/CachedCollection';
+import { CachedCollection } from '../../../../client/lib/cachedCollections/CachedCollection';
 import { CachedChatRoom } from './CachedChatRoom';
 
 declare module '@rocket.chat/core-typings' {
