+++ apps/meteor/app/models/client/models/ChatPermissions.ts
@@ -1,6 +1,6 @@
 import type { IPermission } from '@rocket.chat/core-typings';
 
-import { CachedCollection } from '../../../ui-cached-collection/client';
+import { CachedCollection } from '../../../../client/lib/cachedCollections';
 
 export const AuthzCachedCollection = new CachedCollection<IPermission>({
 	name: 'permissions',
