--- apps/meteor/app/ui-cached-collection/client/index.ts
@@ -1,2 +1,0 @@
-export { CachedCollection } from './models/CachedCollection';
-export { CachedCollectionManager } from './models/CachedCollectionManager';
