+++ apps/meteor/client/importPackages.ts
@@ -42,6 +42,5 @@ import '../app/utils/client';
 import '../app/settings/client';
 import '../app/models/client';
 import '../app/ui-utils/client';
-import '../app/ui-cached-collection/client';
 import '../app/reactions/client';
 import '../app/livechat/client';
