+++ apps/meteor/client/lib/cachedCollections/CachedCollection.ts
@@ -6,12 +6,12 @@ import { Meteor } from 'meteor/meteor';
 import { Mongo } from 'meteor/mongo';
 import { ReactiveVar } from 'meteor/reactive-var';
 
-import type { MinimongoCollection } from '../../../../client/definitions/MinimongoCollection';
-import { baseURI } from '../../../../client/lib/baseURI';
-import { getConfig } from '../../../../client/lib/utils/getConfig';
-import { isTruthy } from '../../../../lib/isTruthy';
-import { withDebouncing } from '../../../../lib/utils/highOrderFunctions';
-import { sdk } from '../../../utils/client/lib/SDKClient';
+import { sdk } from '../../../app/utils/client/lib/SDKClient';
+import { isTruthy } from '../../../lib/isTruthy';
+import { withDebouncing } from '../../../lib/utils/highOrderFunctions';
+import type { MinimongoCollection } from '../../definitions/MinimongoCollection';
+import { baseURI } from '../baseURI';
+import { getConfig } from '../utils/getConfig';
 import { CachedCollectionManager } from './CachedCollectionManager';
 
 export type EventType = 'notify-logged' | 'notify-all' | 'notify-user';
