+++ apps/meteor/client/lib/cachedCollections/index.ts
@@ -1,0 +1,2 @@
+export { CachedCollection } from './CachedCollection';
+export { CachedCollectionManager } from './CachedCollectionManager';
