+++ apps/meteor/client/startup/incomingMessages.ts
@@ -2,8 +2,8 @@ import type { IMessage } from '@rocket.chat/core-typings';
 import { Meteor } from 'meteor/meteor';
 
 import { ChatMessage } from '../../app/models/client';
-import { CachedCollectionManager } from '../../app/ui-cached-collection/client';
 import { sdk } from '../../app/utils/client/lib/SDKClient';
+import { CachedCollectionManager } from '../lib/cachedCollections';
 
 Meteor.startup(() => {
 	Tracker.autorun(() => {
